services/gateway/pyproject.toml
[project]
name = "alfred-gateway"
version = "0.1.0"
description = "Alfred gateway — WhatsApp management service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "uvicorn[standard]>=0.34.0,<1.0.0",
    "sqlmodel>=0.0.24,<0.1.0",
    "pydantic-settings>=2.8.0,<3.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "python-multipart>=0.0.20,<0.1.0",
    "alfred-shared",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
